Ported library to newer versions of mac os

+++ src/mac/LightAquaBlue/BBStreamingOutputStream.m
@@ -34,10 +34,10 @@
     return self;
 }
 
-- (int)write:(const uint8_t *)buffer maxLength:(unsigned int)len
+- (int)write:(const uint8_t *)buffer maxLength:(NSUInteger)len
 {
     //NSLog(@"[BBStreamingOutputStream] writing data...");
-    int buflen = [mDelegate write:[NSData dataWithBytesNoCopy:(void *)buffer 
+    int buflen = [mDelegate write:[NSData dataWithBytesNoCopy:(void *)buffer
                                                        length:len
                                                  freeWhenDone:NO]];
     return buflen;
